Titles for toolbar buttons, fixes #72
Revision to my other pull-request (#121). ESLint issues have been resolved.

This is a fix for the missing tooltips on the toolbar buttons. The button titles
were added to `locales/en-US/notes.properties` in order to be translated for
the supported languages. `panel.js` makes use of `i18n.getMessage()` in order
to fetch titles.