Added a 'close' button for sync-note
This is a fix for issue #115: missing 'x / close' button for the sync-note.
Currently, the only way to close the sync-note is by clicking the sync button
again which is not very intuitive.

These changes add an 'x' (specifically the multiplication sign due to it's
symmetry) in the top right corner of the sync-note and when pressed, toggles
the visibility of the note, causing it to become not visible. Pressing the
sync button will cause the message to reappear.

Originally the 'syncNotes' message was being applied to the `<div id="sync-note">`
using the `.textContent` property. This sets the textual content of the specified
node AND all of its descendants. In order to add a close button, the 'syncNotes'
message needed to be added to not the parent <div> but instead a <p> inside of
it. This allowed for the addition of another <div> which could hold the close
button and not get overwritten by the `.textContent` property.