Updated sync note 'Close' button to use close.svg
Replaced the multiplication sign originally in use for the 'Close' button symbol
with the included 'close.svg' file. Other changes are as follows:

- styles.css: styling changes to align the 'close' symbol in the top-right
corner of the sync note.

- panel.html: replacing the multiplication sign with an `<img>` tag that uses
the 'close.svg' file as it's source.

- panel.js: Created and initialized a `const` for the `#sync-note-dialogue`
element. This specific element is the one that will have it's textContent
set to the `syncNotReady` message.