Sync button styling (hover and active)
Fixes #129. Added styling for the `:hover` and `:active` states. Colors used
are based off the Photon UI visual design guide (1).

The button wasn't actually flush with the bottom of the Notes panel so I
moved it downwards with a negative margin value. I then compensated by adding
the same amount of padding to the bottom of the button to keep the contents
of the button horizontally centered.

1. https://mozilla.invisionapp.com/share/PYC5LJJXG#/screens/229940646