Fixed margin issue with Sync note on different browsers (#138) r=vladikoff
* Fix margin issue with Sync note on diff browsers

Just downloaded Firefox Dev edition and started testing Notes on there. When
opening up the Sync note, I saw that there was a great deal of margin to it's
right (the note was not the full width of the sidebar like the Sync button).

On the regular edition of Firefox, the Sync note **was** the full width of the
sidebar. These small changes make it so that the Sync note is the full width of
the sidebar on both browsers.

* Removed the unnecessary "%"