Revised styling for the Sync note and button
Made the necessary changes to decrease the size of the sync note. Now the sync
note has much less padding than before and looks like the proposed design. And
the sync button also looks like the proposed design with the hover and active
states changing the background color but not taking the full width and height
of the button.