Sync button styling (hover and active) (#133) r=vladikoff
* Sync button styling (hover and active)

Fixes #129. Added styling for the `:hover` and `:active` states. Colors used
are based off the Photon UI visual design guide (1).

The button wasn't actually flush with the bottom of the Notes panel so I
moved it downwards with a negative margin value. I then compensated by adding
the same amount of padding to the bottom of the button to keep the contents
of the button horizontally centered.

1. https://mozilla.invisionapp.com/share/PYC5LJJXG#/screens/229940646

* Revised Sync button styling to match design

Button now has a background color whose size is not the entire width and height
of the button. Per the proposed design, background color on hover and active
states have some padding, leaving the button's original white background still
visible.

* Revised styling for the Sync note and button

Made the necessary changes to decrease the size of the sync note. Now the sync
note has much less padding than before and looks like the proposed design. And
the sync button also looks like the proposed design with the hover and active
states changing the background color but not taking the full width and height
of the button.

* Removed `padding-bottom` from #footer

By removing the `padding-bottom` from the footer, the sync button's hovering
and active background color becomes symmetrical/uniform on all sides.

* Reworked sync button to match proposed design

Quite a few changes were made in order to get the sync button to have rounded
corners with the necessary `:hover` and `:active` states highlighted correctly.

The main changes are as follows:

- an extra `<div>` was added that encapsulates the sync button - this allows
for the `border-radius` and `background-color` changes on hovering and clicking

- the encapsulating `<div>` has a 5px padding which keeps it's white background
visible even when the button is hovered (just like in the designs)

- to account for the padding around the button, the button's contents (icon and
text) was shifted slightly to the left so that the button does not expand when
the sidebar is at it's least width