Theme changes applied upon saving in settings.js
background.js:
When Notes is initially started, the default theme is set in local storage.
There is also a listener for messages from `settings.js` which listens for
the theme change. When this message is received, `background.js` sends a
message to `panel.js` which triggers the theme change immediately.

panel.js:
Added message listener waiting for theme change from `background.js`. Also
added an event listener for 'DOMContentLoaded' that fetches saved theme from
storage which is triggered anytime Notes is closed then reopened and applies
the respective theme.

settings.js:
When new theme is selected and saved, a message is sent to `background.js`.
Also included is an event listener for 'DOMContentLoaded' which fetches the
saved theme from local storage and updates the appropriate radio button.