Refined `styles-dark.css` and application of theme
In `panel.html`, the `styles-dark.css` file is always present, but by default
set to disabled. When the user selects the dark theme in the settings page,
the `<link id="dark-css"...` which contains the `styles-dark.css` becomes
enabled. Selecting the default/light theme sets the dark theme to disabled.

The latest version of `styles-dark.css` only overrides any color or resource
that is specific to the dark theme. This allows for easier maintenance of
styling for Notes: any styling changes can be made in `styles.css` and will be
present in both the light and dark themes. When there is an addition/edit of colors,
only those need to be added to `styles-dark.css`.