Rewrite sync logic to use Kinto
This commit shifts the source of truth to be the Kinto collection,
which at present only has a single record (the singleNote
record). Locally, this record should have a content field which stores
the quill content, as well as an id and a last_modified. This record
is transformed by the remote transformer, which encrypts the whole
record, which is stored under the "content" field of the remote
record. A vague attempt at backwards compatibility is made, where we
recognize old-style encrypted records, which are just the "content"
field encrypted, and not a whole Kinto record, and when we recognize
them, we "decode" them as complete Kinto records.

Although the source of truth is now the Kinto collection, no effort
whatsoever was made to update the rest of this project, which still
relies on storing last_modified and notes in browser.local. The
contentWasSynced field wasn't rethought in any way, and neither was
the syncing strategy (sync on startup and every time the user changes
the text).

The syncing logic was adapted from ExtensionStorageSync.jsm in the
mozilla-central tree. There are lots of edge cases that remain to be
covered, especially centering around merge conflicts. No effort was
made to test the cases where we get 401s or key conflicts, although
the code seems sane.