Add support for links
This is the initial implementation for supporting links (#233). With these changes,
links will automatically be created from anything that begins with `http://`
or `https://`. That includes any typed or pasted text.

The changes in the CSS make it so that the Quill link editing UI tooltip is
no longer shown. In order to open a link, the user must mouse over the link
and right-click for the context menu or `Ctrl/Cmd + Click` to open in a new tab.