Bug fixes for #252, #254, and #255
The issue #252 is resolved by adding a function that adds a matcher to the
clipboard which looks for urls. When pasted, the urls are immediately resolved
as links.

Issues #254 and #255 have been resolved by adding a simple `else` statement to
the function that looks out for the `insert` in the quill Delta. In addition
to the `else` statement, the original `if` statement has been made more specific
to match only when the `Delta.ops` length is equal to two (2).