Convert to karma tests
This seems to be the standard for web extensions.

These can be run using npm.

Karma relies on sinon-chrome, which (by default) exposes its stuff as
a "chrome" variable instead of the "browser" variable we expect. The
background script also uses the storage API immediately, so we need to
make sure the mock is populated with data before we even load it.