Use mocha reporter
Revert "Use mocha spec reporter.", moving from the karma spec reporter
back to the karma mocha reporter, just because it seems better
supported and is part of the example-webextension standard we're
trying to follow. Try to achieve the same effect using the built-in Karma
option reportSlowerThan -- set it to a generous 40ms.

This reverts commit 54299d275c0020d26fdef602a44cb0b46eb94b3f.