Give pasted text any active formatting
This is the second part that fixes #271. In the last PR, typing links
with active formatting was applied, but when pasting links any for-
matting would be dropped.

This PR fixes this issue by getting the active formats at the focus
of the editor then applies the formatting to the clipboard text
before pasting it to the editor. This applies to both text and links as
discussed in #322.