(Tried) Building Classic editor from source
This was an attempt to build the Classic editor from source as outlined
in this guide: https://docs.ckeditor.com/ckeditor5/latest/builds/guide-
s/integration/advanced-setup.html.

This isn't bundled correctly however since ClassicEditor.js and
Markdown.js are at the top-level of the directory and need to be inside
the `sidebar/` directory.