Open Notes before pasting content in notepad
Previously, if a user tried to use the 'Send to Notes' context menu op-
tion when Notes was closed, their selected text would not be pasted in
the notepad due to a connection issue between the background and content
script.

This commit makes it so that the Notes panel opens before trying to send
the selected text to be pasted.