Export: menu item to download Notes content
This is a proof of concept attempt at implementing #189. A menu item
has been added in the material design lite menu with a text context of
"Export". When clicked, a file titled "notes.md" will be downloaded on
to the user's computer. The contents of "notes.md" is the contents of
the Notes notepad, retrieved using the CKEditor5 `editor.getData()`
method. This outputs the HTML of the editor contents - for example,
'Hello World' with no formatting would be '<p>Hello World</p>' in
"notes.md".