Add Turndown as dependency to handle HTML -> MD
Using Turndown, we are able to convert CKEditor's HTML output into true
Markdown syntax. This commit adds Turndown as a dependency and adds a
second "Export" button: now users could choose whether to download the
contents of their notepad as HTML or Markdown.

The use of two buttons in the MDL menu is subject to change and is only
meant to used as a proof of concept.