Send selection text once editor is ready
Earlier after calling `browser.sidebarAction.open()`, a magic number was
used to guess when the sidebar was ready before sending the selection
text. The use of magic numbers has been removed and now we make use of
messages between the content script (Editor.js) and the background.js
script to determine once the editor has been initialized and is ready.
Once this message is received from `background.js`, then the selection
text is sent.