Initial implementation of 'Open Notes in Tab'
This adds a new settings option, "Open Notes in..." which allows users
to choose either to open Notes in the sidebar or in a new tab. The cur-
rent implementation will open Notes in the tab when: 1) the browser ac-
tion button is pressed, or 2) the "Send to Notes" context menu option
is selected. Before opening a new Notes tab, all of the tabs open will
be queried to see if a Notes tab exists - if so, then there will be no
action, making sure there are no duplicate Notes tabs.

The edge cases that will still open / show Notes in the sidebar despite
the "Open Notes in new tab" option being selected are when the keyboard
shortcut is used, or when Notes is manually selected from the dropdown
menu in the sidebar.