Add CD10 to all metrics (#842) r=@vladikoff
Add cd10 to every metrics request. Possible values are: 'error', 'isSyncing', 'synced', 'openLogin', 'verifyAccount', 'reconnectSync', and 'signIn'.
Names are the one used by front-end to display proper footer UI.

We are not able for now to detect if a user synced on more than one device. This should be an improvement while working on the v5.mobile app.

Fix #771