Bump version to 4.3.0dev (#1308)
The latest master has the same version number as our released version.
We want to avoid doing this because it can lead to Firefox's auto-update issue where even though you have the latest master installed, you'll get updated to the Test Pilot release version.

We've had issues in the past that were caused by this. See #332 and #339.